backend/requirements.txt
flask
bcrypt
flask_cors
mysql-connector-python
sentry-sdk